+++ addons/serial/serial.c
@@ -1013,16 +1013,16 @@ int rt_spset_thrs(unsigned int tty, int rxthrs, int txthrs)
  * 		-EINVAL		if wrong parameter value
  *
  */ 
-int rt_spset_callback_fun(unsigned int tty, void (*callback_fun)(int, int), 
+long rt_spset_callback_fun(unsigned int tty, void (*callback_fun)(int, int), 
                           int rxthrs, int txthrs)
 {
-	int prev_callback_fun;
+	long prev_callback_fun;
 
 	CHECK_SPINDX(tty);
 	if (rt_spset_thrs(tty, rxthrs, txthrs)) {
 		return -EINVAL;
 	}
-	prev_callback_fun = (int)spct[tty].callback_fun;
+	prev_callback_fun = (long)spct[tty].callback_fun;
 	spct[tty].callback_fun = callback_fun;
 	return prev_callback_fun;
 }
@@ -1047,12 +1047,12 @@ int rt_spset_callback_fun(unsigned int tty, void (*callback_fun)(int, int),
  * 		-EINVAL		if wrong parameter value
  *
  */ 
-int rt_spset_err_callback_fun(unsigned int tty, void (*err_callback_fun)(int))
+long rt_spset_err_callback_fun(unsigned int tty, void (*err_callback_fun)(int))
 {
-	int prev_err_callback_fun;
+	long prev_err_callback_fun;
 
 	CHECK_SPINDX(tty);
-	prev_err_callback_fun = (int)spct[tty].err_callback_fun;
+	prev_err_callback_fun = (long)spct[tty].err_callback_fun;
 	spct[tty].err_callback_fun = err_callback_fun;
 	return prev_err_callback_fun;
 }
